night commute - some issue about TIME ZONE stamp of current date and data

+++ applicatura_weather_app/Model/JSON/WeatherData.swift
@@ -27,7 +27,7 @@ struct Daily: Codable {
 }
 
 struct Weather: Codable {
-    let id: Float
+    let id: Int16
 }
 
 struct Temp: Codable {
